PG170317A:
Power management for CC1350 begins to work. I have removed the old
and useless option for fixed-size process table. It was getting in the
way of progress, so to speak. I can put the device into sleep and
standby modes, but there still seems to be some (minor) current
leakage. The launchpad is not extremely friendly from the viewpoint of
current measurement, as there is extra circuitry on the PCB with some
potential for leaks. The shutdown mode (next step) should be easy,
as it is basically a power-off mode with reset on defineable pins.

I think it's time to do some documenting.

Oh, added the new version of AP331 to the ALPHATRONICS board set, as
well as an OLIMEX board variant to test their new voltage sensors. The
latter will be be probably removed soon.